Update README headlines
This updates the headlines of the README to be more "in order".

I also moved the section for creating a Scenic application from scratch
further down, as it's not necessarily part of the "setup" of the project
itself (and one does not need to do it if it's just about running the
example).