Remove `$` signs before command line input
I removed all the `$` signs before command line input. I feel
that indicating whether something is input for the terminal is more of a
formulating or stying problem, rather than it should be solved with a
the $ character. For most people in workshops this was usually more
disturbing than helpful (they did not know the convention and also
entered the $...).